Double was converted to float

+++ hybrid_image/amdpool/hybrid_image.h
@@ -23,13 +23,13 @@ using namespace std;
 
 void dut( hls::stream<bit32_t> &strm_in, hls::stream<bit32_t> &strm_out);
 
-void FFT(int dir, long m, complex <double> x[]);
+void FFT(int dir, long m, complex <float> x[]);
 
-void fftshift(complex<double> out[4096], complex<double> in[4096], int xdim, int ydim, int xshift, int yshift);
+void fftshift(complex<float> out[4096], complex<float> in[4096], int xdim, int ydim, int xshift, int yshift);
 
-void GaussFilter(int imgwidth, int imgheight, complex<double> F[4096], bool High);
+void GaussFilter(int imgwidth, int imgheight, complex<float> F[4096], bool High);
 
-void normalize(complex<double> imgNormIn[4096], double imgNormOut[4096]);
+void normalize(complex<float> imgNormIn[4096], float imgNormOut[4096]);
 
-void hybrid_image(int intImgSize, complex<double> imgLo_input[4096], complex<double> imgHi_input[4096],  double imgOutput[4096]);
+void hybrid_image(int intImgSize, complex<float> imgLo_input[4096], complex<float> imgHi_input[4096],  float imgOutput[4096]);
